Describe and Generate Game Using Gradio/Main/requirements.txt
requests
Pillow
python-dotenv
gradio